requirements.txt
flask=2.0.1
Flask-SQLAlchemy=2.5.1
flask-login=0.5.0
Flask-Admin=1.5.8